+++ schema-bson/src/main/scala/zio/blocks/schema/bson/ObjectIdSupport.scala
@@ -26,14 +26,14 @@ import zio.blocks.typeid.{Owner, TypeId}
  *
  * The ObjectId is represented as a wrapper around String (hex representation),
  * but when encoded to BSON, it uses the native BsonType.OBJECT_ID (12-byte
- * format) via special handling in BsonSchemaCodec.
+ * format) via special handling in BsonCodecDeriver.
  */
 object ObjectIdSupport {
 
   /**
    * Schema for org.bson.types.ObjectId.
    *
-   * Manually constructed with typeId for "ObjectId" so BsonSchemaCodec can
+   * Manually constructed with typeId for "ObjectId" so BsonCodecDeriver can
    * detect it and use zio-bson's native BsonCodec[ObjectId] (encodes as
    * BsonType.OBJECT_ID).
    */
